+++ accounts/serializers.py
@@ -1,3 +1,4 @@
+from drf_spectacular.utils import extend_schema_serializer
 from rest_framework import serializers
 from django.conf import settings
 from django.contrib.auth import get_user_model
@@ -8,6 +9,7 @@ from .models import Workspace, WorkspaceWebhook
 
 User = get_user_model()
 
+@extend_schema_serializer(component_name='AccountsUser')
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
         model = User